Make model fields and link export mixins
